requirements.txt
alembic==1.12.1
aniso8601==10.0.0
bcrypt==4.2.1
blinker==1.8.2
click==8.1.8
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==2.1.5
ps==0.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.0.0
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.23
typing_extensions==4.12.2
Werkzeug==2.3.7
zipp==3.20.2
python-dotenv==1.0.0
